--- src/main/java/com/kodilla/ecommercee/GroupNotFoundException.java
@@ -1,4 +1,0 @@
-package com.kodilla.ecommercee;
-
-public class GroupNotFoundException extends Exception{
-}
