+++ src/main/java/com/kodilla/ecommercee/domain/Group.java
@@ -1,0 +1,6 @@
+package com.kodilla.ecommercee.domain;
+
+public class Group {
+    private Long id;
+    private String name;
+}
