+++ src/main/java/com/kodilla/ecommercee/domain/GroupDTO.java
@@ -1,22 +1,13 @@
 package com.kodilla.ecommercee.domain;
 
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
 public class GroupDTO {
     private Long id;
     private String name;
-
-    public GroupDTO(Long id, String name) {
-
-        this.id = id;
-        this.name = name;
-    }
-
-    public Long getId() {
-
-        return id;
-    }
-
-    public String getName() {
-
-        return name;
-    }
 }
